qhouyee/7/improve-role-based-interactions: refactor to parse permissions only once
